0.1.2

+++ src/models/program.rs
@@ -14,7 +14,7 @@
 #[derive(Clone, Debug, PartialEq, Default, Serialize, Deserialize)]
 pub struct Program {
     #[serde(rename = "id")]
-    pub id: i32,
+    pub id: i64,
     #[serde(rename = "eventId")]
     pub event_id: i32,
     #[serde(rename = "serviceId")]
@@ -46,7 +46,7 @@ pub struct Program {
 }
 
 impl Program {
-    pub fn new(id: i32, event_id: i32, service_id: i32, network_id: i32, start_at: i32, duration: i32, is_free: bool) -> Program {
+    pub fn new(id: i64, event_id: i32, service_id: i32, network_id: i32, start_at: i32, duration: i32, is_free: bool) -> Program {
         Program {
             id,
             event_id,
